:sparkles:(lld) add default analytics for deeplinks
