+++ .changeset/wicked-fireants-pay.md
@@ -1,0 +1,5 @@
+---
+"ledger-live-desktop": patch
+---
+
+Add default analytics for deeplinks
